tital in list stream

+++ internal/delivery/http/dto/response/list_response.go
@@ -1,8 +1,8 @@
 package response
 
 type ListReponse[T any] struct {
-	Items  []T `json:"items"`
-	Total  int `json:"total"`
-	Limit  int `json:"limit"`
-	Offset int `json:"offset"`
+	Items  []T   `json:"items"`
+	Total  int64 `json:"total"`
+	Limit  int   `json:"limit"`
+	Offset int   `json:"offset"`
 }
